requirements.txt
streamlit>=1.32
tensorflow>=2.13
numpy>=1.24
pandas>=2.0
Pillow>=9.5
